+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Service/StudentService.java
@@ -13,7 +13,7 @@ public interface StudentService {
 
      void createStudent(StudentForRegisterDTO student) throws Exception;
 
-     void deleteStudent(int studentId);
+     void deleteStudent(Long studentId);
 
      void upgradeStudent(StudentDTO student);
